Fixed clang warnings

+++ src/spokk/image_file.c
@@ -24,8 +24,11 @@
 #define IMAGEFILE__MAX(a, b) ((a) > (b) ? (a) : (b))
 
 static int IsSubresourceValid(const ImageFile *image, const ImageFileSubresource subresource) {
-  return (subresource.mip_level >= 0 && subresource.mip_level < image->mip_levels && subresource.array_layer >= 0 &&
-      subresource.array_layer < image->array_layers);
+  return (
+    (int32_t)subresource.mip_level >= 0 &&
+    subresource.mip_level < image->mip_levels &&
+    (int32_t)subresource.array_layer >= 0 &&
+    subresource.array_layer < image->array_layers);
 }
 
 uint32_t ImageFileGetBytesPerTexelBlock(ImageFileDataFormat format) {
